Implemented updating, publish/unpublish, delete and displaying the posts with most recent on editor side

+++ cakephp/app/views/posts/show_editor_posts.ctp
@@ -13,9 +13,13 @@
                 </div>
                 <div class="panel-footer">
                 	<ul class="nav nav-inline">
-                		<li><a href="/posts/publishPost/<?php echo $post['Post']['id']; ?>">Publish</a></li>
-                		<li><a href="/posts/updatePost/<?php echo $post['Post']['id']; ?>">Edit</a></li>
-                		<li><a href="/posts/deletePost/<?php echo $post['Post']['id']; ?>">Delete</a></li>
+                		<?php if($post['Post']['state'] == "Create" || $post['Post']['state'] == "UnPublish"){ ?>
+                			<li><a href="/posts/publishPost/<?php echo $post['Post']['id']; ?>">Publish</a></li>
+                		<?php }else{ ?>
+                			<li><a href="/posts/unpublishPost/<?php echo $post['Post']['id']; ?>">UnPublish</a></li>
+                		<?php } ?>
+                		<li><a href="/posts/getPostForEdit/<?php echo $post['Post']['id']; ?>">Edit</a></li>
+                		<li><a onclick="return confirm('Are you sure to delete this post?')" href="/posts/deletePost/<?php echo $post['Post']['id']; ?>">Delete</a></li>
                 	</ul>
                 </div>
             </div>
